Extract common code to BaseConditionConstraintAnalyzer
