+++ src/nunit.analyzers/ConstraintUsage/EqualConstraintUsageAnalyzer.cs
@@ -1,20 +1,16 @@
 using System.Collections.Immutable;
-using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
 using NUnit.Analyzers.Constants;
-using NUnit.Analyzers.Extensions;
 using static NUnit.Analyzers.Constants.NunitFrameworkConstants;
 
 namespace NUnit.Analyzers.ConstraintUsage
 {
     [DiagnosticAnalyzer(LanguageNames.CSharp)]
-    public class EqualConstraintUsageAnalyzer : BaseAssertionAnalyzer
+    public class EqualConstraintUsageAnalyzer : BaseConditionConstraintAnalyzer
     {
-        internal const string SuggestedConstraintString = nameof(SuggestedConstraintString);
-
         private static readonly string IsEqualTo = $"{NameOfIs}.{NameOfIsEqualTo}";
         private static readonly string IsNotEqualTo = $"{NameOfIs}.{NameOfIsNot}.{NameOfIsEqualTo}";
 
@@ -26,56 +22,11 @@ namespace NUnit.Analyzers.ConstraintUsage
             defaultSeverity: DiagnosticSeverity.Info,
             description: EqualConstraintUsageConstants.Description);
 
-        private static readonly string[] SupportedPositiveAssertMethods = new[] {
-            NameOfAssertThat,
-            NameOfAssertTrue,
-            NameOfAssertIsTrue
-        };
-
-        private static readonly string[] SupportedNegativeAssertMethods = new[] {
-            NameOfAssertFalse,
-            NameOfAssertIsFalse
-        };
-
         public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => ImmutableArray.Create(descriptor);
 
-        protected override void AnalyzeAssertInvocation(SyntaxNodeAnalysisContext context,
-            InvocationExpressionSyntax assertExpression, IMethodSymbol methodSymbol)
+        protected override (DiagnosticDescriptor descriptor, string suggestedConstraint) GetDiagnosticData(
+            SyntaxNodeAnalysisContext context, ExpressionSyntax actual, bool negated)
         {
-            var negated = false;
-
-            if (SupportedNegativeAssertMethods.Contains(methodSymbol.Name))
-            {
-                negated = true;
-            }
-            else if (!SupportedPositiveAssertMethods.Contains(methodSymbol.Name))
-            {
-                return;
-            }
-
-            var constraint = assertExpression.GetArgumentExpression(methodSymbol, NameOfExpressionParameter);
-
-            // Constraint should be either absent, or Is.True, or Is.False
-            if (constraint != null)
-            {
-                if (!(constraint is MemberAccessExpressionSyntax memberAccess
-                       && memberAccess.Expression is IdentifierNameSyntax identifierSyntax
-                       && identifierSyntax.Identifier.Text == NameOfIs
-                       && (memberAccess.Name.Identifier.Text == NameOfIsTrue
-                           || memberAccess.Name.Identifier.Text == NameOfIsFalse)))
-                {
-                    return;
-                }
-
-                if (memberAccess.Name.Identifier.Text == NameOfIsFalse)
-                {
-                    negated = !negated;
-                }
-            }
-
-            var actual = assertExpression.GetArgumentExpression(methodSymbol, NameOfActualParameter)
-                ?? assertExpression.GetArgumentExpression(methodSymbol, NameOfConditionParameter);
-
             var shouldReport = false;
 
             // 'actual == expected', 'Equals(actual, expected)' or 'actual.Equals(expected)'
@@ -86,11 +37,8 @@ namespace NUnit.Analyzers.ConstraintUsage
                 shouldReport = true;
             }
 
-            // 'actual != expected', '!Equals(actual, expected)' or '!actual.Equals(expected)'
-            else if (actual.IsKind(SyntaxKind.NotEqualsExpression)
-                || (IsPrefixNotExpression(actual, out var notOperand)
-                    && (IsStaticObjectEquals(notOperand, context)
-                        || IsInstanceObjectEquals(notOperand, context))))
+            // 'actual != expected'
+            else if (actual.IsKind(SyntaxKind.NotEqualsExpression))
             {
                 shouldReport = true;
                 negated = !negated;
@@ -99,24 +47,10 @@ namespace NUnit.Analyzers.ConstraintUsage
             if (shouldReport)
             {
                 var suggestedConstraint = negated ? IsNotEqualTo : IsEqualTo;
-                var properties = ImmutableDictionary.Create<string, string>().Add(SuggestedConstraintString, suggestedConstraint);
-                context.ReportDiagnostic(Diagnostic.Create(descriptor, actual.GetLocation(), properties, suggestedConstraint));
+                return (descriptor, suggestedConstraint);
             }
-        }
 
-        private static bool IsPrefixNotExpression(ExpressionSyntax expression, out ExpressionSyntax operand)
-        {
-            if (expression is PrefixUnaryExpressionSyntax unarySyntax
-                && unarySyntax.IsKind(SyntaxKind.LogicalNotExpression))
-            {
-                operand = unarySyntax.Operand;
-                return true;
-            }
-            else
-            {
-                operand = null;
-                return false;
-            }
+            return (null, null);
         }
 
         private static bool IsStaticObjectEquals(ExpressionSyntax expressionSyntax, SyntaxNodeAnalysisContext context)
